+++ docs/src/SelectList.doc.js
@@ -34,12 +34,6 @@ card(
         href: 'basicExample',
       },
       {
-        name: 'idealErrorDirection',
-        type: `?'up' | 'right' | 'down' | 'left'`,
-        description: 'Preferred direction for the error Flyout to open',
-        defaultValue: 'right',
-      },
-      {
         name: 'name',
         type: '?string',
         href: 'basicExample',
@@ -178,7 +172,7 @@ class SelectListExample extends React.Component {
         <SelectList
           id="city"
           name="city"
-          errorMessage="This field can not be blank"
+          errorMessage="This selection has an error"
           onChange={this.handleChange}
           options={cityOptions}
           placeholder="Select city"