+++ docs/src/TextField.doc.js
@@ -37,12 +37,6 @@ card(
         href: 'basicExample',
       },
       {
-        name: 'idealErrorDirection',
-        type: `'up' | 'right' | 'down' | 'left'`,
-        description: 'Preferred direction for the error Flyout to open',
-        defaultValue: 'right',
-      },
-      {
         name: 'name',
         type: 'string',
       },