+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Minor
 
+- SelectList, TextArea, TextField: Remove Flyout error message, use new FormErrorMessage (#486)
 - Icon: Add new video-camera icon (#491)
 - Avatar: Fixed a bug in MS Edge where text was not vertically centered (#492)